+++ src/app/page.tsx
@@ -1,13 +1,11 @@
 import FeatureSection from "@/components/FeatureSection/page";
 import HeroSection from "@/components/HeroSection/page";
-import ContactSection from "@/components/ContactSection/page";
 
 export default function Home() {
   return (
     <main>
       <HeroSection />
       <FeatureSection />
-      <ContactSection />
     </main>
   );
 }
